requirements.txt
fastapi

uvicorn

sentence-transformers

faiss-cpu

PyMuPDF

PyPDF2

pyttsx3

numpy

ollama

vosk>=0.3.45

sounddevice

python-multipart